new class added

+++ src/Java_Programming/StringCharCounting/SpecialCharCount.java
@@ -9,7 +9,7 @@ public class SpecialCharCount {
         String str =sc.nextLine();
         int count = 0;
         for(int i =0; i<str.length();i++){
-            int hash = (int)str.charAt(i);
+            int hash = str.charAt(i);
             boolean b = (hash > 90 && hash < 97) || (hash > 122 && hash < 177);
 //            if((hash>31&&hash<48)||hash>57&&hash<65)||(hash>90&&hash<97)||(hash>122&&hash<177)){
 //                count++;
